remove many ut and .to change to ops.cast

+++ debug/accuracy_tools/msprobe/test/mindspore_ut/free_benchmark/handler/test_ms_base_handler.py
@@ -95,7 +95,7 @@ class TestBaseHandler(unittest.TestCase):
 
             first_tensor = Tensor([1.0, 1.2], dtype=ms.bfloat16)
             second_tensor = Tensor([1.5, 2.0], dtype=ms.bfloat16)
-            target = ops.max(ops.div(second_tensor.to(ms.float32), first_tensor.to(ms.float32)))[0].item()
+            target = ops.max(ops.div(ops.cast(second_tensor, ms.float32), ops.cast(first_tensor, ms.float32)))[0].item()
             ret = self.base_handler.get_endless_norm(first_tensor, second_tensor, abs_tol)
             self.assertEqual(ret, target)
